packages/processors/palagenai/backend/requirements.txt
Flask==3.0.0
Flask-PyMongo==2.3.0
Flask-CORS==4.0.0
PyJWT==2.8.0
Werkzeug==3.0.1
pymongo==4.6.1
google-cloud-vision==3.7.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
requests==2.31.0
python-dotenv==1.0.0
Pillow
beautifulsoup4==4.12.2
selenium==4.16.0
webdriver-manager==4.0.1
#google-generativeai==0.3.0
google-search-results
chrome-lens-py
# Azure Computer Vision
azure-cognitiveservices-vision-computervision==0.9.0
msrest==0.7.1

# Tesseract OCR
pytesseract==0.3.10

# EasyOCR
easyocr==1.7.1

# Claude AI (Anthropic)
anthropic>=0.39.0

# PDF Processing
pdf2image
PyPDF2==3.0.1

# Document Detection and Image Processing
opencv-python>=4.8.0
numpy>=1.26.0

#torch==2.1.0
#torchvision==0.16.0

# MinIO S3 Storage
minio==7.2.0

# NSQ Message Queue
gnsq==1.0.2

# SSH Client for Remote Worker Management
paramiko>=3.0.0

debugpy
docker>=7.0.0

# LangChain and Ollama Integration
langchain==0.1.20
ollama>=0.1.0
websockets>=11.0
websockets
